+++ examples/experimental_average.cpp
@@ -19,7 +19,7 @@ int main() {
     idx N = 10000; // number of "measurement experiments"
     for (idx i = 0; i < N; ++i) {
         auto measured = measure(psi, evects);
-        idx m = std::get<RES>(measured); // measurement result
+        idx m = std::get<measure_idx::res>(measured); // measurement result
         if (evals[m] < 0) {
             --res; // -1
         } else {